+++ lensmemoriae_faces/models/__init__.py
@@ -1,3 +1,4 @@
 from . import lensmemoriae_person
 from . import lensmemoriae_face
 from . import lensmemoriae_image
+from . import res_config_settings